chore: remove custom fields on uninstall (#156)
