+++ lending/install.py
@@ -234,33 +234,3 @@ def after_install():
 	make_property_setter_for_journal_entry()
 	print("\nRunning post-install patches to patch existing data...\n")
 	run_patches(get_post_install_patches())
-
-
-def before_uninstall():
-	delete_custom_fields(LOAN_CUSTOM_FIELDS)
-
-
-def delete_custom_fields(custom_fields):
-	"""
-	:param custom_fields: a dict like `{'Customer': [{fieldname: 'test', ...}]}`
-	"""
-
-	for doctypes, fields in custom_fields.items():
-		if isinstance(fields, dict):
-			# only one field
-			fields = [fields]
-
-		if isinstance(doctypes, str):
-			# only one doctype
-			doctypes = (doctypes,)
-
-		for doctype in doctypes:
-			frappe.db.delete(
-				"Custom Field",
-				{
-					"fieldname": ("in", [field["fieldname"] for field in fields]),
-					"dt": doctype,
-				},
-			)
-
-			frappe.clear_cache(doctype=doctype)
